refactor(exceptions): move exceptions into per-layer exceptions.py modules
Exceptions no longer live buried inside logic modules; each layer that owns
exceptions gets a dedicated exceptions.py (Meaningfy convention):

- loader/domain/exceptions.py: LoadingError, ConfigError, UnsupportedFormatError,
  GraphLoadError, ValidationError
- loader/adapters/exceptions.py: GraphStoreError (subclasses domain LoadingError)
- diffing/domain/exceptions.py: VersionMissing, VersionExists, VersionsDeltaExists
- diffing/adapters/exceptions.py: FusekiException, SubprocessFailure

All importers repointed to the new homes. Verified: ruff clean, mypy clean
(59 files), import-linter 10/10, 236 unit tests pass (6 pre-existing env failures).
