+++ rdf_differ/loader/adapters/graph_store.py
@@ -9,18 +9,6 @@ store lib. The engine-selecting composition helper lives in ``graph_store_provid
 
 from typing import Protocol, runtime_checkable
 
-from rdf_differ.loader.domain.model import LoadingError
-
-
-class GraphStoreError(LoadingError):
-    """A triple-store transport/engine failure (HTTP error, parse error, …).
-
-    ``status`` carries the HTTP status when the failure came from a remote
-    endpoint (used to decide whether a retry is worthwhile); ``None`` otherwise.
-    """
-
-    status: int | None = None
-
 
 @runtime_checkable
 class GraphStorePort(Protocol):
