+++ rdf_differ/loader/domain/model.py
@@ -25,26 +25,7 @@ from pydantic import (
     ValidationError as PydanticValidationError,
 )
 
-
-# --- errors -------------------------------------------------------------------
-class LoadingError(Exception):
-    """Base class for all RDF Loading Module errors."""
-
-
-class ConfigError(LoadingError):
-    """The loading configuration is invalid (caught before any write)."""
-
-
-class UnsupportedFormatError(LoadingError):
-    """A version file uses an unknown/unsupported RDF serialisation."""
-
-
-class GraphLoadError(LoadingError):
-    """A version file could not be parsed/loaded into a graph."""
-
-
-class ValidationError(LoadingError):
-    """Post-load structural or delta-content validation failed."""
+from rdf_differ.loader.domain.exceptions import ConfigError
 
 
 # --- enums --------------------------------------------------------------------
